Instantiate tags on request only
Tags are now requested by the template rather than passed into it, so
we're not instantiating tags for no reason. getDefaultTemplateTags()
just returns an array of tag handle/class name pairs, rather than the
actual instances, and the instantiating happens in getTemplateTag(),
which accepts a tag handle and returns the tag instance.