Add a 'redirectLoginPageIfLoggedIn' config setting.
Only affects front-end requests, but if you're logged in and and don't have CP access permissions and hit the loginPath path, determins where you'll be redirected.  By default it is '', which will return you to the home page.

If you do have CP access, you'll get redirected to the dashboard, still.