Changed the behavior when the user doesn't have offline access
If the user is logged in, rather than getting auto-logged out, they now get a 503 error with a clear message and a Log Out link.

If the user is not logged in, and it's a CP request, they will get redirected to the Login page. Otherwise they'll get a standard 503 error.