Several image quality fixes
Fix a bug where if you had defaultImageQuality or a transform image quality set to 0, you'd get a PHP error.

Fix a bug where image quality settings for PNG's would have the opposite affect of what you would expect (like 0-100, higher is better for JPGs).  For PNGs, 100 would be the worst (use most compression).

Not that it matters, because PNG compression is worthless for both Imagick and GD.