Revert "Add uiTimestamp and author name to "current" revision entry dropdown."
This reverts commit 918763ccef753d8c8edcce7671ec46553a550454.

This info is misleading, as all other revisions show the *creator* name (not the same as the entry's official author) and the *save date* (not the same as the entry's official post date). If we want to show similiar information for "Current", we need to be fetching the latest revision of the entry and pulling the creator/save date info from that.