Fixed another user permission bug
If a user was assigned a permission that was nested under one of the permissions they inheretid from a group, and then either they lost the group assignment or the group lost the permission, the user account wolud be in an awkward state where they still had the nested permission that had been assigned directly, but not the permission(s) leading up to it. (At least until their user account was re-edited.)

Now, any permissions a user inherits from their group will be directly assigned to the user as well, in the event that one of its nested permissions had been assigned directly to the user.