Fixed ElementEditor locale permissions bug
It was possible to edit elements in locales the user didn't have permission to edit if the ElementEditor was loaded for an element that was being displayed in that non-editable locale.

Now the PHP checks to make sure the user can edit the current locale; if not it will find the first locale that is supported by the element which the user does have permission to edit.