Add `url` option to DB config
PaaS solutions often provide database credentials in the form of a single string via an environment variable. This string contains the driver, username, password, server, port and database name.

This patch defines the `url` option in the config, parses the `url` option parts if defined, and then assigns the parts to existing variables when the DB config is loaded, having no impact on existing calls to the DB config.