Update Dutch translation for "asset"
Throughout Craft, the Dutch word 'Materiaal' is used wherever the English noun 'Asset' is shown. 'Materiaal' is word that, although it CAN mean 'the material needed to fill a site', doesn't quite fit here. In my opinion, 'Bestand' would be better, and the plural 'Bestanden' is already used in the navigation sidebar. 'Bestand' literally means 'File'.