Stop relying on CONVERT_TZ()
MySQL might not be configured with timezone data (https://stackoverflow.com/q/14454304/1688568), and even if it is. it doesn't support UTC+13:45 or UTC+14:00 timezones.

We can't just normalize dates to UTC and keep getting aggregate counts using a single query, because you without CONVERT_TZ(), some values would end up shifting to the wrong day. So multiple queries it is.

Fixes #2434