@web in Valet causes issues
Hi

English is not my primary language so please change according. Also another user used variables instead of alias so perhaps that's better.

When I first ran into this issue it took me like 5 hours to figure out why I couldn't run multisite in Valet. I know a few other users have had this issue as I helped them out with it. So I thought it was about time I proposed a change in the docs to include a tip.