Updates Schema.php to support pdo_mysql.default_socket
Updates Schema.php to support pdo_mysql.default_socket for Unix socket usage when using mysqldump. The current code does not look for the default PDO socket path, which results in mysqldump failing to connect to the socket if mysqldump itself is not explicitly told (via MySQL configuration or similar) where the socket exists. For server installations that use the standard pdo_mysql.default_socket INI setting, this is unnecessary since that default socket location can be looked up and attempted. That's what this code change does.