Include solution for pagination
I had problems caching my loop with pagination until I saw this post: https://craftcms.stackexchange.com/questions/26950/cache-not-invalidating-when-using-pagination-with-cache?rq=1

Before that my cache was not invalidated when saving a new entry.

Feel free to use it and to change the text and placement. I just know it could be useful to have there as other than me has run into this "issue".