[WIP] Add complete support for plugins including when they are at root
Currently the implementation of adding plugins is a concept. I cant imagine this is the best way of doing doing this.

Aside from this rudimentary project config support is added. The client can now define a project config file which will be added.

See also my commerce fork for this system in action on the commerce tests