Add doesTableExist MigrationHelper function
When uninstalling a plugin, sometimes we might try dropping foreign keys on a
table that doesn't exist yet. This function lets us check whether or not
that table exists yet before dropping foreign keys.

For example, in `Install::safeDown()` we could have this which would throw if the table hadn't been created yet:

```php
MigrationHelper::dropAllForeignKeysToTable('{{%table_does_not_exist}}', $this);
$this->dropTableIfExists('{{%table_does_not_exist}}');
```

With this, we can check if the table exists first:

```php
if (MigrationHelper::doesTableExist('{{%table_does_not_exist}}', $this) {
    MigrationHelper::dropAllForeignKeysToTable('{{%table_does_not_exist}}', $this);
    $this->dropTable('{{%table_does_not_exist}}');
}
```