Docs.
My first attempt at basic documentation for the Craft module.

It has:
An intro page
A getting started guide
An overview of the config parameters the Craft module requires/offers
Some slight extra info on plugins and modules.
Pages with TODO: Tags which can act as a roadmap
Edited the .vuepress config.