[WIP] Implement a system  for console commands.
The way it works in practice is similar to https://laravel.com/docs/5.8/console-tests

It features a fluid-interface that allows you to define the steps a console-command must enter through directly in accordance with yii\console\Controller.

If any steps dont match it errors out.

Currently only  stdOut is mapped but the concept works as expected and mapping the other Yii2 console output methods aren't difficult.

See craftunit\console\ConsoleTest on an implementation

Docs still need to be writted