@angrybrad Slightly improved b838f0089543b538614c21937b9c5dc30202884f.
If we were ever to write a test that required searching for the user that was created during the test setup in `install.php`. This would delete the reference to this user.

We used to do this globally in the Craft module but removed it for cleanliness (https://github.com/craftcms/cms/blob/a2f55a3266cd8096c999213e68fa3eb1a03803d0/src/test/Craft.php#L188)