Adding new checks to old migrations scripts for core craft upgrades.  The following tables were already included in the Install.php script and thus were causing an error during the DB Migration Process:
  Table::DRAFTS
  Table::REVISIONS
  Table::GQLSCHEMAS
  Table::SEQUENCES
  {{%entryversionerrors}}
  {{%entrydrafterrors}}

Steps to reproduce:
You will need to trigger one of the affected migrations. this entails rolling back to a snapshot from a previous craft version or removing the respective entry from the craft migration table.
With your composer version pointed to the latest craft release, Remove the entire vendor directory and run a fresh composer install.
Go to the /admin install page and click 'Finish Up'.
You see migration run for awhile before breaking on the error "craft_xxxx" table already exists.

Why this should work:
It is a standard pattern in the migration scripts to run `dropTableIfExists` on a table before running `createTable`.  This will allow th e migration process to continue for tables that are created as part of Install.php.