Fixed #10184
Numeric strings passed to `Formatter::asDecimal()` (etc.) could be deterined to be misrepresented when formatted. E.g. there is no way to format "123abc" as a number accurately. When that is the case, the fallback formatter will be used, which doesn't support the locale's custom number formats. So the value could end up with the wrong thousands separator/decimal symbols, but still thought to be formatted for the locale.

Two part fix:

- Added `willBeMisrepresented()` to the formatter. When that returns `true`, we just output the number verbatim and don't pass the `[locale]` value (so Craft knows it wasn't formatted).
- Default/Min/Max settings are now normalized to an int/float if possible, rather than saved as strings. So less chance of being considered misrepresented in the first place.