Properly calculate height of transformed images
Fixes an issue where the height of a non-cropped image being transformed was being calculated as a percentage of the height rather than of the width.

#11837

Fixes an issue where the height of a non-cropped image being transformed was being calculated as a percentage of the height rather than of the width.

#11837