Better loading of globals twig extension
Previously it was possible for Craft to attempt loading the `GlobalsExtension()` before Craft was installed. If the developer had custom error templates, this would result in a default "internal server error" message rather than rendering the custom error template. This fixes that by checking to see if Craft is installed before loading the `GlobalsExtension()`.