Adjust asset upload process
Previously if the last asset you were uploading threw an error the asset index view would not update because the code responsible for displaying prompts after the last upload was only attached to a successful upload.

This refactors that code a bit so the code responsible for displaying prompts runs after each upload regardless of if they were successful or not.