Fixed a bug where dynamic subfolders would be reported as missing when updating asset indexes
When a folder was created dynamically, only the record was created and not the directory. For S3 this worked just fine because it doesn't really believe in directories, but without the directory being created, it wouldn't get returned in the directory listing from the filesystem and would be considered missing.