Don't factor `$context` in to cache key
Previously we were serializing `$context` into the cache key for graphql queries. While this was good for uniqueness it could make the caches _too_ unique. When the cache is first being generated, fields are collected and memoized which altered the `$context` variable. When two queries were run right after each other, the first cache would miss, generate the cache, alter the context and cause the second query to miss the cache as well.

This solves that by ignoring the `$context` all together when generating the `$cacheKey`. The `$cacheKey` is already made up of a hash of the query, serialized variables and the operation name. The `$context` can't be altered without a plugin so it seems overkill to include within the `$cacheKey`.

Worth noting that this could cause some confusing situations in development if a developer is changing the context via module or plugin but doesn't change the query. With this change and caching enabled they would get stale results.