Initial support for “fill” transform parameter
Adds an additional `fill` parameter to image transforms.

The value of `fill` must be a value that the `ColorValidator` class accepts as valid (right now that means either 3 or 6 digit hex

The fill is only applied when it is valid, the mode is set to `fit` and the upscale images setting is set to `false` OR when the ratio of the transform differs from the ratio of the image itself. In that scenario, blank pixels between the size of the source image and the specified transform will be filled with the color specified (using the image’s color space or palette).

Indexes in the `imagetransformindex` table will now have the fill appended to the end of the string.

Generated folders will also have the fill color appended (minus the `#` character).