Initial component system
Establishes the bones of the component system. The majority of this is heavily inspired by [Symfony’s twig components](https://symfony.com/doc/current/ux-twig-component/index.html) but there are a few Yii-isms in here. The biggest change is throwing everything into a big service class vs a separate factory and renderer.