Matrix field handle validation
Currently, matrix handle length validation doesn't include the block
type handle in the length, this makes it possible for you to create an
overall field handle that's longer than the DB will allow. When this
occurs things get left in a bad state, where some fields are saved and
others fail.

This fixes that issue by adding validation to matrix field handles to
make sure they're not too long when including the block type handle in
the field name.

Fixes #12422