Ensure transform filename reflects format
When the url for an asset is determined within an immediate transform the url is [determined from the image transform index](https://github.com/craftcms/cms/blob/409a40742eb651b63106b78099befffc69b44231/src/imagetransforms/ImageTransformer.php#L87) which happens before the transform is generated, and the filename is set on the index. This results in an incorrect URL being returned the first time `getUrl` is called.