Add Control Panel URL support to UrlHelper
The UrlHelper methods, urlWithScheme and urlWithToken, were updated to include an argument to specify whether or not they are to be used for Control Panel URLs. The testUrlWithScheme function in the UrlHelperTest file was also updated to reflect this change. This makes the urlWithScheme function more flexible in handling different types of URLs.